Changed the .js.applescript extension to .js
The files in the src/ directory cannot be opened with the Mac Script
Editor app any more because of the dependency system.  Therefore,
considering that the purpose of the ".applescript" extension was to
allow editing with the script editor (if a user/developer wished to do
so for some reason), the extension is not longer required.